pyproject.toml
[project]
name = "fastapiproject"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "alembic>=1.17.2",
    "asyncpg>=0.31.0",
    "bcrypt>=4.0.0",
    "fastapi[standard]>=0.127.0",
    "httpx>=0.28.1",
    "polyfactory>=3.2.0",
    "psycopg2-binary>=2.9.11",
    "pwdlib[argon2]>=0.3.0",
    "pydantic-settings>=2.12.0",
    "pydantic[email]>=2.12.5",
    "pyjwt>=2.10.1",
    "python-multipart>=0.0.21",
    "sqlalchemy>=2.0.45",
    "tzdata>=2025.2",
    "uvicorn[standard]>=0.40.0",
]

[dependency-groups]
dev = [
    "aiosqlite>=0.22.1",
    "factory-boy>=3.3.3",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.10",
    "taskipy>=1.14.1",
]

# -------------------------------------------------------
# testes CONFIGURATION
[tool.pytest.ini_options]
pythonpath = "."



# -------------------------------------------------------
# linters CONFIGURATION

[tool.ruff]
line-length = 79

[tool.ruff.format]
preview = true


# -------------------------------------------------------
# taskipy CONFIGURATION


[tool.taskipy.tasks]
lint = 'ruff check'
pre_format = 'ruff check --fix'
format = 'ruff format'
pre_test = 'task lint'
test = 'pytest -s -x --cov=. -vv'
post_test = 'coverage html'
